split petition of transaction into file
