+++ src/petition_types/mod.rs
@@ -1,9 +1,11 @@
 mod builder_entity;
 mod builder_factors_types;
 mod factor_list_kind;
+mod petition_of_transaction;
 mod petitions;
 
 pub use builder_entity::*;
 pub use builder_factors_types::*;
 pub use factor_list_kind::*;
+pub(crate) use petition_of_transaction::*;
 pub(crate) use petitions::*;
